user-service/requirements.txt
Django==5.2.12
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
requests==2.31.0
